refactor(js): remove values folder
