+++ app/javascript/angular/tests/session.test.js
@@ -1,5 +1,5 @@
 import test from "blue-tape";
-import * as Session from "../../javascript/values/session";
+import * as Session from "../../javascript/session";
 
 test("when title is missing it defaults to unnamed", t => {
   const session = { selectedStream: {} };
