+++ app/javascript/javascript/drawSession.js
@@ -1,7 +1,7 @@
 import _ from "underscore";
 const note = process.env.NODE_ENV === "test" ? "" : require("./note");
 const drawNotes = note.drawNotes;
-import * as Session from "./values/session";
+import * as Session from "./session";
 import { locationMarkersByLevel } from "./theme";
 import heat from "./heat";
 import map from "./map";
