+++ app/javascript/javascript/session.js
